% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambms-package.R
\docType{package}
\name{gambms-package}
\alias{gambms-package}
\alias{_PACKAGE}
\title{gambms: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
