% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambmsmethods.R
\name{gambms}
\alias{gambms}
\title{Title}
\usage{
gambms(
  fm,
  dat,
  knotConfig = c("EVEN", "VS", "FREE"),
  prior = c("Unit", "Hyper-g", "Uniform", "Hyper-g/n", "Beta-prime", "ZS-adapted",
    "Robust", "Intrinsic", "constant", "CH"),
  family = c("poisson", "gaussian", "bernoulli"),
  Ctrl = list()
)
}
\arguments{
\item{Ctrl}{}
}
\description{
Title
}
