% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncs.R
\name{ncs}
\alias{ncs}
\title{A function to construct a natural cubic spline basis function}
\usage{
ncs(x, nk = 10)
}
\arguments{
\item{x}{variable}

\item{nk}{integer; number of knots for each variable}
}
\value{
ncsSmoothSpec object
}
\description{
for internal usage
}
\seealso{
\code{\link[=gambms]{gambms()}}
}
