% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambmsmethods.R
\name{plotnumknot}
\alias{plotnumknot}
\title{The function that plots a posterior distribution of the number of knots}
\usage{
plotnumknot(x, n_row, ...)
}
\arguments{
\item{x}{gambms object;}

\item{n_row}{integer; a number of rows in a plot}

\item{...}{}
}
\value{
none
}
\description{
The function that plots a posterior distribution of the number of knots
}
