% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambmsmethods.R
\name{print.gambms}
\alias{print.gambms}
\title{The function that prints a gambms object}
\usage{
\method{print}{gambms}(x)
}
\arguments{
\item{x}{gambms S3 object;}

\item{...}{}
}
\value{
none
}
\description{
The function that prints a gambms object
}
