% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalers.R
\name{scalefrom01}
\alias{scalefrom01}
\title{Scale back to the original scale}
\usage{
scalefrom01(X, mins = NULL, maxs = NULL)
}
\arguments{
\item{X}{A matrix to be scaled column-wise.}

\item{mins}{A vector of minimums of scaled intervals. Its length should be equal to the number of columns of \code{X}.}

\item{maxs}{A vector of maximums of scaled intervals. Its length should be equal to the number of columns of \code{X}.}
}
\value{
A scaled matrix.
}
\description{
Scale back to the original scale
}
