% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmat.R
\name{simmat}
\alias{simmat}
\title{A function to simulate data frame}
\usage{
simmat(
  flist,
  xmin = 0,
  xmax = 1,
  n = 1000,
  family = c("bernoulli", "poisson", "gaussian"),
  link = NULL,
  sig = NULL,
  intercept = 0,
  beta = NULL
)
}
\arguments{
\item{flist}{list; a list of true (uncentered) functions}

\item{xmin}{real; support of x}

\item{xmax}{real; support of x}

\item{n}{integer; number of samples}

\item{family}{character; "bernoulli", "poisson", "gaussian"}

\item{link}{character; "log", "logit", "identity"}

\item{sig}{real; standard deviation of gaussian}

\item{intercept}{real; intercept}

\item{beta}{numeric vector; coefficients of linear terms, if included}
}
\value{
data frame
}
\description{
A function to simulate data frame
}
