% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambmsmethods.R
\name{summary.gambms}
\alias{summary.gambms}
\title{The function that summarizes a gambms object}
\usage{
\method{summary}{gambms}(x)
}
\arguments{
\item{x}{gambms S3 object;}

\item{...}{}
}
\value{
none
}
\description{
The function that summarizes a gambms object
}
