% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambms-package.R
\docType{package}
\name{gambms-package}
\alias{gambms-package}
\title{gambms: Model Selection-based Estimation for Generalized Additive Models with Mixtures of g-priors}
\description{
This package provides RCpp implementations of estimation methods for generalized additive models based on Bayesian model selection with mixtures of g-priors.
}
\author{
\strong{Maintainer}: Gyeonghun Kang \email{hunlearning@yonsei.ac.kr}

}
\keyword{internal}
