% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambmsmethods.R
\name{plotresiduals.gambms}
\alias{plotresiduals.gambms}
\title{Title}
\usage{
\method{plotresiduals}{gambms}(fit, nReplications = 10, nBins = NULL, ylim, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
