% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmat.R
\name{simmat}
\alias{simmat}
\title{Title}
\usage{
simmat(
  flist,
  xmin = 0,
  xmax = 1,
  n = 1000,
  family = c("bernoulli", "poisson", "gaussian"),
  link = NULL,
  sig = NULL,
  intercept = 0,
  beta = NULL
)
}
\arguments{
\item{beta}{}
}
\description{
Title
}
